+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 
 All notable changes to this project will be documented in this file.
+## [1.0.0]() (2025-07-03)
+
+### Features
+
+* Add `enabled_notification` to enable notification which send messages through webhook to Slack
+* Add `slack_webhook_url` add slack webhook which will handle messages
 
 ## [0.1.22]() (2024-12-05)
 
@@ -8,7 +14,6 @@ All notable changes to this project will be documented in this file.
 
 * Add `var.enable_backend` to enable backend or use frontend format only
 * Add `aws_amplify_webhook` to provides an Amplify Webhook resource
-*
 
 ## [0.1.4]() (2024-12-05)